Move authorization helpers to their own concern.
